HTMLFormElement changed to SelectElement on PremisesSelector component
